refactor(user): move logout mutation hook to feature hooks directory
The logout mutation hook was moved from the generic api directory to the user feature hooks directory to better organize code by feature. The old implementation was removed and the import in Settings.tsx was updated to use the new location.
